requirements.txt
bitarray==1.2.1
pycryptodome==3.9.7
